+++ src/engine/btree.ts
@@ -2,9 +2,8 @@
 
 import { BiMap } from "mnemonist";
 import copy from "fast-copy";
-import { printDebug, config, BTreeInsertionPoint } from "../util";
+import { printDebug, InsertionPoint } from "../util";
 import * as Engine from "./common";
-import { workspace } from "../index";
 
 /*
  * Notes about BTree -
@@ -65,6 +64,7 @@ class RootNode extends TreeNode {
 }
 
 export class TilingEngine implements Engine.TilingEngine {
+    settings: Engine.Settings = new Engine.Settings;
     // turn the desktop into a string first so its indexed by primitive instead of reference
     rootNode: RootNode = new RootNode;
     // changed when desktop is changed
@@ -251,18 +251,21 @@ export class TilingEngine implements Engine.TilingEngine {
     }
     
     addClient(client: KWin.AbstractClient): boolean {
+        if (this.settings.insertionPoint == InsertionPoint.Active) {
+            const lastClient = this.settings.lastActiveClient;
+            if (lastClient != null && lastClient.tile != null) { // or undefined
+                const tile = lastClient.tile;
+                if (this.nodeMap.inverse.has(tile) && tile.parent != null) return this.putClientInTile(client, tile);
+            }
+        }
         // truly this is the peak of programming
         let stack: Array<TreeNode> = [this.rootNode];
         let stackNext: Array<TreeNode> = [];
-        const targetClient = workspace.previousActiveClient
         let i = 0;
         stackloop: while (stack.length > 0) {
             for (const node of stack) {
                 if (node.children == null) {
                     if (node.client != null) { // case for basically all non-root tiles
-                        if (config.btreeInsertionPoint == BTreeInsertionPoint.Active && targetClient && node.client != targetClient) {
-                            continue;
-                        }
                         node.split();
                         node.children![0].client = node.client;
                         node.children![1].client = client;
@@ -279,7 +282,7 @@ export class TilingEngine implements Engine.TilingEngine {
             }
             stack = stackNext;
             // invert insertion order every 2 iterations if option is enabled to put windows on right
-            if (config.btreeInsertionPoint == BTreeInsertionPoint.Right && i % 2 == 0) {
+            if (this.settings.insertionPoint == InsertionPoint.Right && i % 2 == 0) {
                 stack.reverse();
             }
             stackNext = [];