+++ changelog.md
@@ -5,6 +5,8 @@
 ### 0.6.0
 * Sort of support minimum window sizes
 * Tile resizing done in pixels instead of relatives
+* Added live settings configuration per desktop
+* Added insertion points to Half and ThreeColumn
 
 ## 0.5